# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Phonebook app",
    "version": "1.0.0",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is the official API documentation for the Phonebook app. The API allows users to manage their contacts, including adding, editing, and deleting them.\nThe authentication system supports secure login and registration, including token-based authorization.\nThis documentation follows the OpenAPI 3.1.0 standard and can be previewed using Redocly or Swagger-UI.\n"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    },
    {
      "url": "phonebook-app-0rap.onrender.com"
    }
  ],
  "tags": [
    {
      "name": "Contacts",
      "description": "Manage contacts in the phonebook, including creation, update, deletion, and retrieval."
    },
    {
      "name": "Auth",
      "description": "User authentication and session management, supporting login, registration, and token-based authorization."
    }
  ],
  "paths": {
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Retrieve a paginated list of contacts",
        "operationId": "getStudents",
        "description": "Returns a paginated list of contacts belonging to the authenticated user. Supports filtering by contact type and favourite status, sorting by any contact field, and pagination via query parameters.\n\n  This route requires a valid access token (bearerAuth).\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "perPage",
            "description": "Number of contacts per page",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "example": "name",
              "description": "Field name to sort by (e.g. name, contactType, etc). Default is `_id`"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "description": "Sort direction: ascending (`asc`) or descending (`desc`)",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "example": "asc"
            }
          },
          {
            "in": "query",
            "name": "contactType",
            "description": "Filter contacts by category",
            "schema": {
              "type": "string",
              "enum": ["work", "home", "personal"],
              "example": "personal"
            }
          },
          {
            "in": "query",
            "name": "isFavourite",
            "description": "Filter contacts by \"favourite\" status",
            "schema": {
              "type": "boolean",
              "example": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved list of contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contacts!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/contact"
                          }
                        },
                        "page": {
                          "type": "number",
                          "example": 2
                        },
                        "perPage": {
                          "type": "number",
                          "example": 10
                        },
                        "totalItems": {
                          "type": "number",
                          "example": 47
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 5
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": true
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Create a new contact",
        "operationId": "createContact",
        "description": "Create a new contact and adds it to the user's personal contact list. This endpoint supports optional email and photo fields, and requires a valid access token (bearerAuth).\n\n  The request must be sent as `multipart/form-data` to enable optional photo upload.\n  The `contactType` must be one of: `work`, `home`, or `personal`.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "phoneNumber", "contactType"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Full name of the contact",
                    "example": "John Doe"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Phone number of the contact",
                    "example": "+380631234567"
                  },
                  "email": {
                    "type": ["string", "null"],
                    "description": "Email address of the contact (optional)",
                    "example": "john@example.com"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Whether the contact is marked as favourite",
                    "example": true,
                    "default": false
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Category of the contact",
                    "enum": ["work", "home", "personal"],
                    "example": "personal"
                  },
                  "photo": {
                    "type": ["string", "null"],
                    "format": "binary",
                    "description": "Contact photo file (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a new contact",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a contact!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Retrieve a contact by ID",
        "operationId": "getContactById",
        "description": "Returns a contact that matches the provided unique identifier (ID). The endpoint is secured and requires a valid access token (bearerAuth). If the contact does not exist or does not belong to the authenticated user, a `404 Not Found` response will be returned.\n\n  You can use this endpoint to view full details of a contact,\n  including name, email, phone number, and photo (if available).\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Unique identifier of the contact (MongoDB ObjectId)",
            "schema": {
              "type": "string",
              "example": "64f7e05a0c0e5568fda154c8"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved contact",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contact with id 64f7e05a0c0e5568fda154c8"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "tags": ["Contacts"],
        "summary": "Update an existing contact by ID",
        "operationId": "updateContactById",
        "description": "Updates one or more fields of an existing contact that belongs to the authenticated user. The request must be sent as `multipart/form-data` to support optional photo updates.\n\n  Only provided fields will be updated; omitted fields will remain unchanged.\n  A valid access token (bearerAuth) is required.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Unique identifier of the contact (MongoDB ObjectId)",
            "schema": {
              "type": "string",
              "example": "64f7e05a0c0e5568fda154c8"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Full name of the contact",
                    "example": "John Doe"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Phone number of the contact",
                    "example": "+380631234567"
                  },
                  "email": {
                    "type": ["string", "null"],
                    "description": "Email address of the contact (optional)",
                    "example": "john@example.com"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Whether the contact is marked as favourite",
                    "example": true,
                    "default": false
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Category of the contact",
                    "enum": ["work", "home", "personal"],
                    "example": "personal"
                  },
                  "photo": {
                    "type": ["string", "null"],
                    "format": "binary",
                    "description": "Contact photo file (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated a contact",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully patched a contact!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/contact"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete an existing contact by ID",
        "operationId": "deleteContactById",
        "description": "Deletes a contact that belongs to the authenticated user by the specified ID.\n  Requires a valid access token (bearerAuth).\n\n  If the contact does not exist or is not accessible by the user,\n  a `404 Not Found` response will be returned.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Unique identifier of the contact (MongoDB ObjectId)",
            "schema": {
              "type": "string",
              "example": "64f7e05a0c0e5568fda154c8"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Contact deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 204
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contact": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "phoneNumber",
          "contactType",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique identifier of the contact",
            "example": "64f7e05a0c0e5568fda154c8"
          },
          "name": {
            "type": "string",
            "description": "Full name of the contact",
            "example": "John Doe"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number of the contact",
            "example": "+380631234567"
          },
          "email": {
            "type": ["string", "null"],
            "description": "Email address of the contact (optional)",
            "example": "john@example.com"
          },
          "isFavourite": {
            "type": "boolean",
            "description": "Whether the contact is marked as favourite",
            "example": true,
            "default": false
          },
          "contactType": {
            "type": "string",
            "description": "Category of the contact",
            "enum": ["work", "home", "personal"],
            "example": "personal"
          },
          "userId": {
            "type": "string",
            "description": "ID of the user who owns this contact",
            "example": "65bf1234a4e145ff6d3edc10"
          },
          "photo": {
            "type": ["string", "null"],
            "format": "uri",
            "description": "URL to the contact's photo if available",
            "example": "https://res.cloudinary.com/demo/image/upload/sample.jpg"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of contact creation",
            "example": "2025-05-17T10:23:00.123Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of last contact update",
            "example": "2025-05-18T14:45:27.456Z"
          }
        }
      }
    },
    "responses": {
      "401": {
        "description": "Unauthorized – access token missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message", "data"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Contact not found"
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    }
  }
}